refactor: migrate consumers to ontology.Service
Wires internal/ontology into pipeline, doctor, audit, onboarding, and
tui. Removes the interim build-tag split:

- pipeline/executor: gains an ontology field, constructs Service via
  ontology.New(Config{Enabled: EnabledFromManifest(m)}, Deps{...});
  replaces maybeWarnOntologyStale/buildOntologySection/recordOntologyUsage
  with Service method calls. Deletes pipeline/ontology_{enabled,disabled}.go.
- audit.Logger: drops LogOntologyInject/Lineage/Warn; adds generic
  LogEvent(kind, body). Ontology-specific formatting now lives in the
  ontology package and writes via the AuditSink interface.
- doctor: extracts checkOntology into checks_ontology.go; checks_config
  delegates.
- onboarding/ontology_step: single untagged implementation; install of
  the post-merge hook delegates to ontology.InstallStalenessHook.
- tui/ontology_provider: reads staleness via ontology.IsStaleInDir,
  removing the build-tag file-pair.
